# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Princess Empire E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResult {
  token: string;
  role: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  images: string[];
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  sizes?: string[];
  colors?: string[];
  stock: number;
  /** @nullable */
  brand?: string | null;
  featured?: boolean;
  bestSeller?: boolean;
  flashSale?: boolean;
  /** @nullable */
  flashSaleEndsAt?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  sizes?: string[];
  colors?: string[];
  stock: number;
  /** @nullable */
  brand?: string | null;
  featured?: boolean;
  bestSeller?: boolean;
  flashSale?: boolean;
  /** @nullable */
  flashSaleEndsAt?: string | null;
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  oldPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  sizes?: string[];
  colors?: string[];
  stock?: number;
  /** @nullable */
  brand?: string | null;
  featured?: boolean;
  bestSeller?: boolean;
  flashSale?: boolean;
  /** @nullable */
  flashSaleEndsAt?: string | null;
}

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  productCount?: number | null;
}

export interface CategoryInput {
  name: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface OrderItem {
  productId: number;
  name: string;
  price: number;
  quantity: number;
  image: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  items: OrderItem[];
  /** @nullable */
  notes?: string | null;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrdersPage {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface StoreSettings {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  storeName?: string | null;
  /** @nullable */
  storeEmail?: string | null;
  /** @nullable */
  storeAddress?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
}

export interface StoreSettingsInput {
  whatsappNumber: string;
  /** @nullable */
  storeName?: string | null;
  /** @nullable */
  storeEmail?: string | null;
  /** @nullable */
  storeAddress?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
}

export type DashboardStatsTopCategoriesItem = {
  name: string;
  count: number;
};

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  recentOrders: Order[];
  topCategories: DashboardStatsTopCategoriesItem[];
  lowStockProducts?: Product[];
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  sort?: ListProductsSort;
  /**
   * @nullable
   */
  featured?: boolean | null;
  /**
   * @nullable
   */
  bestSeller?: boolean | null;
  /**
   * @nullable
   */
  flashSale?: boolean | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  brand?: string | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
};

export type ListProductsSort =
  | (typeof ListProductsSort)[keyof typeof ListProductsSort]
  | null;

export const ListProductsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  popular: "popular",
} as const;

export type GetRecentProductsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetFeaturedProductsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetBestSellersParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetFlashSalesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetSearchSuggestionsParams = {
  q: string;
};

export type ListOrdersParams = {
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};