# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Princess Empire E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersPage"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        productCount:
          type: ["integer", "null"]
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [productId, name, price, quantity, image]
      properties:
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        image:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, customerName, customerEmail, customerPhone, address, items, total, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, address, items]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: ["string", "null"]
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
    OrdersPage:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...