# lib/api-zod/src/generated/types/ordersPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Princess Empire E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export interface OrdersPage {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}